refactor: consolidate calendar constants into lib/constants.ts
