+++ client/src/lib/constants.ts
@@ -14,3 +14,20 @@ export const PERIOD_DURATION_MIN = 1;
 export const PERIOD_DURATION_MAX = 10;
 export const CONTRACEPTION_MAX = 100;
 export const NOTES_MAX = 500;
+
+// calendar
+export const CALENDAR_CONFIG = {
+  showNeighboringMonth: true,
+  showFixedNumberOfWeeks: false,
+  prevLabel: '‹',
+  nextLabel: '›',
+  prev2Label: '«',
+  next2Label: '»',
+  view: 'month' as const,
+  maxDetail: 'month' as const,
+} as const;
+
+export const WEEKDAY_FORMAT = {
+  length: 2,
+  uppercase: true,
+} as const;
